GUI - SWIPE implemented.
Note that if user starts movement from main panel, and at the center screen and goes a little to one side and then returns with some speed, releasing the screen after passing the middle of the screen, then the panel of the opposite side than first showed is shown. This behaviour is not an anomaly or malfunction since other applications have the same behaviour, outstanding Google Play Store.

ATTENTION!: Background service and broadcast receivers have been disabled for battery saving purposes while testing in real devices.

         2014-02-21T21:58:50A